requirements.txt
Flask
flask-cors
roboflow
Pillow
matplotlib
base64